requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
zipp==3.19.2
